token auth for deleting

+++ controllers/blogs.js
@@ -18,20 +18,16 @@ blogsRouter.get('/', async (request, response, next) => {
 })
 
 
-
 blogsRouter.post('/', async (request, response, next) => {
   const body = request.body
-
-  console.log(request)
 
   const decodedToken = jwt.verify(request.token, process.env.SECRET)
   if(!request.token || !decodedToken.id) {
     return response.status(401).json({ error: 'token missing or invalid' })
+
   }
-  console.log(decodedToken)
 
   const user = await User.findById(decodedToken.id)
-  console.log(user)
 
   const blog = new Blog({
     title: body.title,
@@ -54,13 +50,31 @@ blogsRouter.post('/', async (request, response, next) => {
 
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+  const blog = await Blog.findById(request.params.id)
+
+  if(!request.token) {
+    return response.status(401).json({ error: 'token missing' })
+  }
 
   try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if(!decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const userId = await User.findById(decodedToken.id)
+
+    if (blog.user.toString() === userId._id.toString()) {
+      await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    } else {
+      return response.status(401).json({ error: 'you do not have the rigth permission' })
+    }
+
   } catch (error) {
     next(error)
   }
+
 })
 
 blogsRouter.put('/:id', async (request, response, next) => {
